Fix extra message on startup bug
Fix the bug in which an additional message would be unexpectedly
processed on startup. This was due to a misunderstanding of the Slack
realtime API. Occasionally, it sends a copy of your last message if
you're reconnecting after what it thinks is a network failure.
Apparently this is so you know that the message was received properly.
Note that this also explains why reproduction was a bit inconsistent.