Adjust pipe closures in slackio
This change fixes the pipe I/O issues revealed by panicking on Wait
errors.

Basically, the problem is that I was thinking of Go's io.Pipe like a
UNIX pipe, where I've learned that both ends need to be closed at some
point. As a result of this I was blindly closing both ends without
giving it too much thought.

Once I read through the io.Pipe source code, this became much clearer.
Closing **only the write end** of the pipe causes the next read to
return EOF. Actually, this is how UNIX pipes behave as well, so really
the problem is that I'm just stupid in general.

Anyways, in both cases closing the write end of the pipe only is the
correct thing to do. In the reader case, it ensures that the exec
package's goroutine receives EOF and terminates. In the writer case, it
ensures that the internal scanner receives EOF and terminates. And since
the "pipe" isn't managed by the kernel it's not like I need to worry
about returning any resources. It's really just a synchronized buffer of
sorts.

This behavior is sort of implicit in the Go documentation. Maybe I
needed to read that more closely.