Update comments/styling; export SlackIO
Update a bunch of the comments as well as some of the code styling. Most
notably, errors from io.PipeWriter closure are no longer checked since
the Close method of that implementation is defined and *documented* to
always return nil. (I might still bring this back since it causes
gometalinter to complain.)

Also notably, the SlackIO struct is now exported, so that its
documentation comments actually can be visible and make sense. Really,
it's just good style for this type to be exported.