Block slackio init on successful connection
Ensure that we are successfully connected to Slack before allowing other
methods of slackio.Client to proceed. This is useful when running Close
very shortly after connection (even though this is not a good
slackbridge use case). More importantly, this helps if it takes multiple
attempts to connect to Slack.