Add new slackio.Client for RTM communication
Create the new slackio.Client, which encapsulates real-time
communication logic rather than provide a full ReadWriteCloser for Slack
channels. Other parts of slackio will be refactored to depend on this.

This also includes a basic set of test cases, just for critical parts
that I thought could use it. In particular, GetMessageStream is complex
enough that I wanted to make sure it got hit.