Clean up interval batcher termination
Previously, the interval batcher would terminate as soon as an error was
received, potentially even before the output channel was properly
closed. This is against the Batcher contract, which states that
consumers of a batcher should effectively range over its output channel,
then consume the emitted error.

This change fixes up the interval batcher to only stop when the upstream
output channel is closed. Once that's done, it will *then* emit the
upstream error on its output channel.

As long as you don't emit an error on an unbuffered channel before
closing the output, you should be golden.