Minor style/doc updates for batchers
Output flushes and channel closures have been moved out of defer
statements, to make the expected control flow of a batcher more clear.
That is: it closes the output channel, then emits an error, then closes
the error channel. When batchers and their consumers both assume this
order of operations, deadlocks and other problems can be prevented.

Explicit documentation was also added regarding error channel buffering.
While this was implicit in the use of the word "should" regarding error
channel reading, I wanted to make sure this was very clear.