Improve Reader/Client resource cleanup
Ensure that both the case of closing a Reader and the case of Client
closing a message stream channel are properly handled, and that internal
goroutines always get cleaned up as expected.

The main goal of this was to fix up Reader so that it actually drains
the message stream while it is in the process of closing, as
GetMessageStream requires. However, a case that hadn't yet been
considered too closely was that of GetMessageStream closing the channel
on its own. The resolution was to force the caller of GetMessageStream
to close the done channel and unblock Client's internal cleanup
goroutine.

While this clear ownership is good, there is the problem that the stream
channel could now get closed twice if Client is the one that initiates
the closure. I still need to fix this.