Improve Client.Close behavior
Specifically, ensure that there is only ever one call to close each
message stream channel, to prevent a possible panic. Also, Close now
blocks until all channels created with GetMessageStream have
successfully been closed. That is, after Close is called it is
guaranteed that consumers will not receive any further messages.

This is all accomplished by leveraging the existing "done" channel in
each Client. When it is closed, this implicitly causes closure of each
message stream exactly as if a consumer has requested its closure.
Special thanks to the "select" statement for letting that race happen.