Split out slackio from slackbridge
This commit adds a README, Travis CI config, and other basic files to
instantiate slackio as a separate repository from the slackbridge CLI
program.

This change is intended to make slackio more easily usable as an
external package, and to allow things like vendoring of dependencies in
slackbridge itself.

There is some work left to be done to make slackio more robust. Most
notably, I'd like to remove all potential panic sources at some point.
While this may be appropriate enough for slackbridge at this time, it's
not necessarily appropriate for other consumers.