Make Client's auth panic an error rather than a string
This is intended to ensure consistency with the rest of slackio, which
typically panics with error values when necessary.

Note that the panic on a *blank* API token is left as a string, as this
panic is recoverable and could be depended on by consumers. This is
inconsistent with Writer's panic on blank channel IDs, which is an error
value. This is a good candidate for cleanup in a future breaking release
of slackio.