Fix incorrect t.Fatal calls in client tests
I've been running the staticcheck tool (https://staticcheck.io) on
various projects of mine. It noticed that I was using t.Fatal
incorrectly in several of my subscription-related unit tests, as I was
potentially calling it from goroutines other than that running the test
function. As the Go documentation explains, this is incorrect as it only
terminates the current goroutine, not the actual test goroutine or other
goroutines spawned by the test.