Clean up client unit tests
The Subscribe[At] and Unsubscribe operations have now been refactored
into separate test functions, as a comment implied might be done in the
future.

Also, the checking of default subscription IDs was moved out of the
goroutine in which the subscription was created, to reduce the amount of
explicit synchronization / implementation dependence in the tests.